mtd: spi-nand: Return number of corrected bitflips when reading
This commit fixes a bug in the current SPI NAND implementation, which
misses the update of the 'ecc_stats.corrected' field in the MTD main structure.

This field holds the number of corrected ECC bitflips, as reported by
the device's internal ECC controller.

Without this patch, the bitflips aren't reported to upper layers such
as UBI. UBI uses the number of corrected bitflips to decide if an
eraseblock needs scrubbing.

Change-Id: I392cadae5323f5cb78a67284d3ad7f62408c564a
Signed-off-by: Ezequiel Garcia <ezequiel.garcia@imgtec.com>
Reviewed-by: Daniel Ehrenberg <dehrenberg@chromium.org>