mtd: spi-nand: Rework SPI NAND command transfer
Using a separate struct spi_transfer for the command and the address
results in random I/O errors. This commit reworks the way SPI messages
are prepared for each command, putting the command and the address
in a single struct spi_transfer.

While here, add some sanity checks on the command.

Change-Id: Ib6d5f4a9a14ea26d4fb12c20989e28e49eef950e
Signed-off-by: Ezequiel Garcia <ezequiel.garcia@imgtec.com>
Reviewed-by: Daniel Ehrenberg <dehrenberg@chromium.org>