i2c: img-scb: enable fencing for all versions of the ip since 2.2.1
The code to read from the master read fifo, and write to the master
write fifo, checks a bit in an SCB register before every byte to
ensure that the fifo is not full (write fifo) or empty (read fifo).
Due to clock domain crossing inside the SCB block the updated value
of this bit is only visible after 2 cycles.

The scb_wr_rd_fence() function does 2 dummy writes (to the read-only
revision register), and it's called before reading from or writing to the
fifos to ensure that subsequent reads of the fifo status bits do not read
stale values.

The 2 dummy writes are required in all versions of the ip since 2.2.1

Change-Id: I701237984a0d1e1beb8e751045b95185c006556d
Signed-off-by: Sifan Naeem <sifan.naeem@imgtec.com>