pistachio: pll: Fix 32bit integer overflows
This commit fixes 32bit integer overflows throughout the pll driver
(i.e. wherever the result of integer multiplication may exceed the
range of u32).

One of the functions affected by this problem is .recalc_rate. It
returns incorrect rate for some pll settings (not for all though)
which in turn results in the incorrect rate setup of pll's child
clocks.

Change-Id: Ic3c528533d2ce6ddeb17f8f8450ca39018cde8b0
Signed-off-by: Zdenko Pulitika <zdenko.pulitika@imgtec.com>