i2c: img-scb: verify support for requested bit rate
The requested bit rate can be outside the range supported by the driver.
The maximum bit rate this driver supports at the moment is 400Khz.

Return -EINVAL if the bit rate is larger than 400khz.

Maximum speed supported by the driver can be increased to 1Mhz by
adding support for "fast plus mode" in the future.

Change-Id: I12235c5644d08b550a1c0588a4afa41736a4db9d
Signed-off-by: Sifan Naeem <sifan.naeem@imgtec.com>