i2c: img-scb: fix LOW and HIGH period values for the SCL clock
After determining the minimum value for the High period (TCKH) the
remainder of the internal clock pulses is set as the Low period (TCKL).
This causes the i2c clock duty cycle to be much less than 50%.

The fix suggested here, start with TCKH and TCKL at 50% of the internal
clock pulses and adjusts the TCKH and TCKL values from there if the
minimum value for TCKL is not met. This will make sure the i2c clock
duty cycle is at 50% or close 50% whenever possible.

Change-Id: I61f371855becedef058ce5184e2fcb301a099a56
Signed-off-by: Sifan Naeem <sifan.naeem@imgtec.com>