i2c: img-scb: calls to img_i2c_transaction_halt relocated to support and control repeated starts on IP version 3.3.
In version 3.3 of the IP when transaction halt is set, an interrupt
will be generated after each byte of a transfer instead of after
every transfer but before the stop bit.
Due to this behaviour we have to be careful that every time we
release the transaction halt we have to re-enable it straight away
so that we only process a single byte, not doing so will result in
all remaining bytes been processed and a stop bit being issued,
which will prevent us having a repeated start.

This change will have no effect on earlier versions of the IP.

Change-Id: I6b523bc03fe7aec438a4278d11e547eb7bdc4dd0
Signed-off-by: Sifan Naeem <sifan.naeem@imgtec.com>