i2c: img-scb: add handle for stop detected interrupt
Triggered when a Stop bit is detected on the bus. No special
requirement on this interrupt but is just an indicator of end
of current transfer.
Stop bit indicates the end of a transfer and if the transfer
has finished before all data is written to the fifo or with more
data expected return error with transfer complete signal.

Change-Id: Ifc29f0a227555b55fa7d54e5072a424a997492cb
Signed-off-by: Sifan Naeem <sifan.naeem@imgtec.com>