KIN-746: cc2520: add support for extclock-freq configuration
cc2520 support generation of external clock on GPIOn pins. This
patch adds an optional 'extclock-freq' property in device tree for
configuring the frequency.

cc2520 supports a limited number of frequencies, please check
datasheet for the exact value to be set.
extclock is disabled if this property is set to 0, if not specified
extclock is set to 1MHz which is the default value of cc2520 after reset.

Change-Id: I327b8af13a66d20babc86578d93dfea1fcd32cfc
Signed-off-by: Shraddha Chaudhari <Shraddha.Chaudhari@imgtec.com>