pistachio_concerto.dts: cleanup
There is no separate audio board for concerto - it is part of the board and
therefore does not deserve a separate .dtsi file.

NOTE:
Currently there is a mismatch between the definitions of Concerto (.dtsi)
and Concerto mBuB (.dts) - leading to a misuse of pistachio_concerto.dts
where in fact an mBuB is used. This will be addressed in a next changeset.

Change-Id: Ieac2f49aa40ade678844a5f58f3750c86f084ade
Signed-off-by: Marcin Nowakowski <marcin.nowakowski@imgtec.com>