pistachio: hostport: poll for ACK
Relying on an interrupt for every acknowledgement required use of a
semaphore in poking routines, which influenced client's design. They could
only call them from contexts that could sleep. Now img-hostport polls
for the ACK after sending a message which makes it possible
to replace the semaphore with a spinlock.

BUG=none
TEST='hciconfig -a'

Change-Id: I64d3a5c10bfe290c0c1018690b47178ad024ab78
Signed-off-by: Bartosz Flis <bartosz.flis@imgtec.com>