dts: remove NAND boot specific dts files
Before this change, the NAND boot specific dts files would
specify NAND as the location of the the root file system.
These are no longer necessary as now the bootargs are
passed from U-boot and U-boot will configure the location
of the root file system. If that's not done, USB will be
used by default as it's currently stated in the pistachio_*.dts
files.

Change-Id: Iff30bed6b0bbf65e6c7f19a214d434633c3c169b
Signed-off-by: Ionela Voinescu <ionela.voinescu@imgtec.com>