atu: clean up code atu_clk_maintainer.c
The atu_rate_change flag is used to let the application know that a rate change
occurred so that the application will take care of convergence fast. So this
flag is mainly used while adjustimg time/freq. The time of day is now set
regardless of if a rate change has occurred, whereas previously it would not
set the time of day if any rate_change event handling is pending.

Change-Id: I32b019b7ce71b70f8632e68c956b35df84608ad7
Signed-off-by: Surendra Yeragam <Surendra.Yeragamreddy@imgtec.com>