rtc: pistachio: add a rtc driver based on timer1 block.
Add a simple rtc driver based on pistachio timer1 block.
This driver uses time-counter and cycle-counter infrastructure
in kernel to manage the time-stamp value.

Current Timer overflow register will clear on read.
So we cannot use the same timer for alarm so use timer2 block
for alarm interface.

*Note* :
overflow is a 4 bit value giving details
on number of times the 32 bit counter overflowed
if overflow reaches max(0xf) it raises a overflow irq.

Change-Id: Ib1e8a71b8edf76810b52f1d1be527cf7ff2890c2
Signed-off-by: Govindraj Raja <Govindraj.Raja@imgtec.com>