KIN-1414: dts: Pistachio: Remove audio card dts examples
The dts examples are meant for internal reference only.
They are not going to be upstreamed. They also contain examples for hacks which
cannot be upstreamed; for example, the breakout board includes a split i2s example.

Not removing pistachio_bub_audio_no_daughterboard_example.dtsi though, because it
is required for audio on BUBs.

Change-Id: I936d377d290658e4fab978da17c35489d381279c
Signed-off-by: Matin Momin <Matin.Momin@imgtec.com>