dtsi: correct mmc clock speed for pistachio
The sdhost interface on pistachio is designed for up to 50MHz, but
currently the DT entry allows up to 200MHz.

In addition to possible timing violations, A side effect occurs which artificially
slows down cards requesting certain frequencies.  For example, due to the source PLL
running at 350MHz, the frequency passed into the sdhost block is as close as possible
to 200MHz, which is 350/2 = 175MHz.  So a card requesting 50MHz actually receives
175/4=43.75MHz.  This change will pass 350/7 = 50MHz to the block, and 50MHz cards will
now run at 50/1 = 50MHz.

Change-Id: I33532b98f23d1f927aadfb6585860823ed115aec
Signed-off-by: James Hartley <james.hartley@imgtec.com>