KIN-1534: new pistachio_prodtest_defconfig
The prodtest SW needs to use ttyS1 for comms but we currently
use that for the console. Attempted to disable this using
boot args but kernel then resorts to its default of ttyS0
which is used by BT. So instead have created a prodtest kernel
config that has console access to uart disabled completley.

Change-Id: I83ba0311999bff813592c01af2c01dbb0a76f9e8
Signed-off-by: Ian Pozella <Ian.Pozella@imgtec.com>